Worked on getting the javascript/html/css side of things set up so our wasm module can actually get loaded and tested. Put IS_TARGET_WEB around various things that call the javascript functions since they won't be the same when compiling for Orca. Still have a problem with converting C string pointers into javascript pointers. The byte buffer thing is always returning 0 so it thinks all strings are 0 length. Committing so I can pull on my laptop and work on the problem tomorrow

+++ wasm/std/include/internal/wasm_std_common.h
@@ -45,9 +45,9 @@ Description:
 #error This standard library implementation assumes little-endian byte order
 #endif
 
-#define WASM_MEMORY_PAGE_SIZE      (64*1024UL) //64kB or 65,536b
-#define WASM_MEMORY_MAX_NUM_PAGES  (64*1024UL) //65,536 pages * 64 kB/page = 4GB
-#define WASM_MEMORY_MAX_SIZE       (WASM_MEMORY_MAX_NUM_PAGES * WASM_MEMORY_PAGE_SIZE)
+#define WASM_MEMORY_PAGE_SIZE      (64*1024ULL) //64kB or 65,536b
+#define WASM_MEMORY_MAX_NUM_PAGES  (64*1024ULL) //65,536 pages * 64 kB/page = 4GB
+#define WASM_MEMORY_MAX_SIZE       ((u64)WASM_MEMORY_MAX_NUM_PAGES * (u64)WASM_MEMORY_PAGE_SIZE)
 #define WASM_PROTECTED_SIZE        1024       //1kB at start of wasm memory should be unused and should never be written to
 
 #if LANGUAGE_CPP
